// apps/web/src/core/ui/dropdown/types.ts
import type { Button } from '@/core/ui';
import type { ComponentProps, PropsWithChildren } from 'react';

export type DropdownPosition = 'bottomLeft' | 'bottomRight' | 'topLeft' | 'topRight';

export type DropdownProps = PropsWithChildren<{
  isOpen?: boolean;
  onOpenChange?: (isOpen: boolean) => void;
  className?: string;
}>;

export type DropdownTriggerProps = PropsWithChildren<ComponentProps<typeof Button>>;

export type DropdownContentProps = PropsWithChildren<{
  className?: string;
  contentClassName?: string;
  position?: DropdownPosition;
}>;

export type DropdownItemProps = PropsWithChildren<ComponentProps<typeof Button>>;